Apply some more changes from error handling refactor, check if user feedback enabled
