+++ src/SentryServiceProvider.php
@@ -14,8 +14,8 @@ namespace FoF\Sentry;
 use Flarum\Frontend\Assets;
 use Flarum\Frontend\Compiler\Source\SourceCollector;
 use Flarum\Foundation\ErrorHandling\Reporter;
-use Flarum\Foundation\ErrorHandling\ViewRenderer;
-use FoF\Sentry\Renderers\SentryRenderer;
+use Flarum\Foundation\ErrorHandling\ViewFormatter;
+use FoF\Sentry\Formatters\SentryFormatter;
 use FoF\Sentry\Reporters\SentryReporter;
 use Illuminate\Support\ServiceProvider;
 use Sentry\ClientBuilder;
@@ -69,8 +69,8 @@ class SentryServiceProvider extends ServiceProvider
 
         $this->app->alias('sentry', HubInterface::class);
 
-        $this->app->extend(ViewRenderer::class, function (ViewRenderer $renderer) {
-            return new SentryRenderer($renderer);
+        $this->app->extend(ViewFormatter::class, function (ViewFormatter $formatter) {
+            return new SentryFormatter($formatter);
         });
 
         $this->app->tag(SentryReporter::class, Reporter::class);
